Update automator.js
This actually tells how many active raining gold abilities are in place; So I propose checking which one is greater. Downside is this can still earn less gold if the boss doesn't have the most active raining gold abilities. If someone wants to add a check for a boss with active raining gold abilities, I believe that would be optimal.